feat(args): unify arg-parser and support ref/commit trace for vcs/fpga
This change moves the common parts of args struct definition and arg_parser
from emu into shared common directory, ensuring all env initailized with same
default args configuration. And we sperate verilator-specified code from
common arg-parser, leaving it in emu.

For now, Verilator and FPGA use the software-side common arg_parser, while
VCS and Palladium overwrite selected fields via Verilog DPI-C.

Additionally, both VCS and FPGA now support dump-ref-trace and dump-commit-trace,
advancing debugging on software.
